requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
asyncpg
alembic
python-decouple
bcrypt
passlib
python-jose
python-multipart